path: integrate path as a module
also remove unused path functions
