+++ applications/cup2019/path.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <kernel_defines.h>
 
 #include "path.h"
 #include "app.h"
@@ -370,74 +371,6 @@ path_t robot_path = {
     /* static cfg */
     .current_pose_idx = 0,
     .play_in_loop = FALSE,
-    .nb_pose = sizeof(poses) / sizeof(path_pose_t),
+    .nb_poses = ARRAY_SIZE(poses),
     .poses = poses,
 };
-
-inline const path_pose_t *path_get_current_path_pos(const path_t *path)
-{
-    return &path->poses[path->current_pose_idx];
-}
-
-inline const path_pose_t *path_get_pose_at_idx(path_t *path, uint8_t idx)
-{
-    if ((idx < path->nb_pose) && (idx > 0)) {
-        return &path->poses[idx];
-    }
-    else {
-        return NULL;
-    }
-}
-
-inline uint8_t path_get_current_pose_idx(path_t *path)
-{
-    return path->current_pose_idx;
-}
-
-inline void path_set_current_pose_idx(path_t *path, uint8_t idx)
-{
-    path->current_pose_idx = idx;
-}
-
-inline void path_reset_current_pose_idx(path_t *path)
-{
-    path->current_pose_idx = 0;
-}
-
-inline void path_increment_current_pose_idx(path_t *path)
-{
-    if (path->current_pose_idx < path->nb_pose - 1) {
-        path->current_pose_idx += 1;
-    }
-    else if (path->play_in_loop) {
-        path->current_pose_idx = 0;
-    }
-}
-
-inline void path_decrement_current_pose_idx(path_t *path)
-{
-    if (path->current_pose_idx > 0) {
-        path->current_pose_idx -= 1;
-    }
-    else if (path->play_in_loop) {
-        path->current_pose_idx = path->nb_pose - 1;
-    }
-}
-
-inline uint8_t path_get_current_max_speed(const path_t *path)
-{
-    const path_pose_t *current_path_pos = &path->poses[path->current_pose_idx];
-
-    return MIN(current_path_pos->max_speed, MAX_SPEED);
-}
-
-void path_horizontal_mirror_all_pos(const path_t *path)
-{
-    for (int i = 0; i < path->nb_pose; i++) {
-        pose_t *pos = (pose_t *)&path->poses[i].pos;
-
-        pos->x *= -1;
-        pos->O = 180 - pos->O;
-        pos->O = ((int)pos->O) % 360;
-    }
-}
